Extract toggleAllItems to its own file
