+++ src/pages/css-props/index.js
@@ -1,6 +1,7 @@
 import { useMemo } from "react";
 import cssPropsHtml from "./_cssPropsHtml";
 import useHashDetails from "../../components/useHashDetails";
+import { toggleAllItems } from "../../components/toggleAllItems";
 
 function CssProperties() {
   // This prevents *other* sections closing when the URL hash changes on hash-link clicks and the page re-renders:
@@ -12,26 +13,6 @@ function CssProperties() {
   );
 
   useHashDetails();
-
-  const toggleAllItems = (event, listId, toggleId) => {
-    const propListEl = document.getElementById(listId);
-    const allDetails = propListEl.getElementsByTagName("details");
-    const toggleLink = document.getElementById(toggleId);
-    if (toggleLink.hasAttribute("open") == false) {
-      for (let i = 0; i < allDetails.length; i++) {
-        allDetails[i].setAttribute("open", "");
-        toggleLink.setAttribute("open", "");
-      }
-    } else if (toggleLink.hasAttribute("open") == true) {
-      for (let i = 0; i < allDetails.length; i++) {
-        allDetails[i].removeAttribute("open");
-        toggleLink.removeAttribute("open");
-      }
-    }
-    event.preventDefault();
-
-    return false;
-  };
 
   const toggleAllCSS = (event) =>
     toggleAllItems(event, "prop-list", "toggle-css");
